requirements.txt
aerich==0.2.3
aiomysql==0.0.20
aiosqlite==0.15.0
anyio==2.0.0
asyncclick==7.1.2.1
asyncpg==0.21.0
cffi==1.14.2
cryptography==3.1
databases==0.3.2
DateTime==4.3
fastapi==0.61.1
h11==0.10.0
h2==3.2.0
hpack==3.0.0
Hypercorn==0.10.2
hyperframe==5.2.0
idna==2.10
iso8601==0.1.13
passlib==1.7.2
priority==1.3.0
pycparser==2.20
pydantic==1.6.1
PyMySQL==0.9.2
PyPika==0.39.1
pytz==2020.1
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
tortoise-orm==0.16.15
typing==3.7.4.3
typing-extensions==3.7.4.3
uuid==1.30
wsproto==0.15.0
zope.interface==5.1.0
